requirements.txt
asgiref==3.5.0
black==22.1.0
click==8.0.4
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-extensions==3.1.5
django-heroku==0.3.1
django-seed==0.3.1
djangorestframework==3.13.1
Faker==13.3.0
flake8==4.0.1
gunicorn==20.1.0
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
toposort==1.7
typing_extensions==4.1.1
whitenoise==6.0.0
